--- crates/rpc/src/tests/mod.rs
@@ -1,6 +1,0 @@
-#[cfg(test)]
-mod meter;
-#[cfg(test)]
-mod rpc;
-#[cfg(test)]
-mod utils;
